+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All Notable changes to `flipboxdigital\salesforce` will be documented in this file
 
-## 1.0.0
-- Initial release!
+## 3.1.0
+### Added
+- `\Flipbox\Salesforce\Query\DynamicQueryBuilderInterface` to define getting and setting of variables
+
+## 3.0.0
+- Reduce footprint